Move NoteData to notedata.rs and create stub for note manager
